pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "taboo-arena"
version = "0.1.0"
description = "Local Streamlit app for Taboo-like LLM benchmarking."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" },
]
dependencies = [
  "accelerate>=1.6.0",
  "huggingface_hub>=0.30.0",
  "jinja2>=3.1.4",
  "llama-cpp-python>=0.3.5",
  "nltk>=3.9",
  "pandas>=2.2.3",
  "platformdirs>=4.3.6",
  "psutil>=6.1.0",
  "pyarrow>=19.0.0",
  "pydantic>=2.10.0",
  "pyyaml>=6.0.2",
  "rapidfuzz>=3.10.1",
  "streamlit>=1.44.0",
  "torch>=2.6.0",
  "transformers>=4.51.0",
  "typer>=0.15.2",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.15.0",
  "pytest>=8.3.5",
  "ruff>=0.11.2",
]

[project.scripts]
taboo-arena = "taboo_arena.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/taboo_arena"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "B",
  "UP",
]
ignore = [
  "B008",
  "E501",
]

[tool.ruff.lint.isort]
known-first-party = ["taboo_arena"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
mypy_path = "src"
strict = false
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = false
disable_error_code = ["import-untyped"]
plugins = []
